rm CommonStorageEnumsTest
