+++ api/src/test/java/org/pmiops/workbench/db/model/StorageEnumsTest.java
@@ -1,37 +1,57 @@
 package org.pmiops.workbench.db.model;
 
 import static com.google.common.truth.Truth.assertThat;
+import static com.google.common.truth.Truth.assertWithMessage;
 
 import com.google.common.collect.BiMap;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.junit.Test;
+import org.pmiops.workbench.model.Domain;
 
 public class StorageEnumsTest {
   private final Object INDICATES_STATIC_METHOD = null;
 
   final Set<Class> enumClasses = getEnumerationClasses();
-  final Collection<Method> methods = Arrays.asList(DbStorageEnums.class.getDeclaredMethods());
+  final Set<Method> methods =
+      Stream.of(
+              DbStorageEnums.class.getDeclaredMethods(),
+              CommonStorageEnums.class.getDeclaredMethods())
+          .flatMap(Stream::of)
+          .collect(Collectors.toSet());
 
   // e.g. public static Short reviewStatusToStorage(ReviewStatus s)
   final Map<Class, Method> enumClassToStorageMethod =
       methods.stream()
+          // domainToDomainId is stringly typed - test with test_domainId
+          .filter(m -> !m.getName().equals("domainToDomainId"))
           .filter(m -> enumClasses.contains(firstMethodParameterType(m)))
           .collect(Collectors.toMap(this::firstMethodParameterType, m -> m));
 
   // e.g. public static ReviewStatus reviewStatusFromStorage(Short s)
   final Map<Type, Method> storageMethodToEnumClass =
       methods.stream()
+          // domainIdToDomain is stringly typed - test with test_domainId
+          .filter(m -> !m.getName().equals("domainIdToDomain"))
           .filter(m -> enumClasses.contains(m.getReturnType()))
           .collect(Collectors.toMap(Method::getReturnType, m -> m));
+
+  // special-case test for domain ID round trip
+
+  @Test
+  public void test_domainId() {
+    for (Domain v : Domain.values()) {
+      String storageValue = CommonStorageEnums.domainToDomainId(v);
+      assertWithMessage("unmapped enum value: " + v).that(storageValue).isNotNull();
+      assertThat(v).isEqualTo(CommonStorageEnums.domainIdToDomain(storageValue));
+    }
+  }
 
   @Test
   public void test_noMissingMapEntries() throws InvocationTargetException, IllegalAccessException {
@@ -58,15 +78,17 @@ public class StorageEnumsTest {
   }
 
   /**
-   * Retrieve all Enum classes in DbStorageEnums
+   * Retrieve all Enum classes in DbStorageEnums and CommonStorageEnums
    *
-   * <p>our convention for BiMaps in DbStorageEnums is <Enum, Short> so we retrieve the first
-   * parameterized type
+   * <p>our convention for BiMaps in DbStorageEnums/CommonStorageEnums is <Enum, Short> so we
+   * retrieve the first parameterized type
    *
    * @return
    */
   private Set<Class> getEnumerationClasses() {
-    return Stream.of(DbStorageEnums.class.getDeclaredFields())
+    return Stream.of(
+            DbStorageEnums.class.getDeclaredFields(), CommonStorageEnums.class.getDeclaredFields())
+        .flatMap(Stream::of)
         .filter(field -> field.getType().equals(BiMap.class))
         .map(
             field -> {
